src/rq3/code_search/cloud_compute/CodeScanning/requirements.txt
google-cloud-storage
pydriller==1.15.1
regex